+++ src/main/java/io/github/devatherock/artifactory/entities/DockerLayer.java
@@ -1,5 +1,6 @@
 package io.github.devatherock.artifactory.entities;
 
+import io.micronaut.core.annotation.Introspected;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,6 +8,7 @@ import lombok.Setter;
 
 @Getter
 @Setter
+@Introspected
 @NoArgsConstructor
 @AllArgsConstructor
 public class DockerLayer {