+++ src/main/java/io/github/devatherock/logback/LogbackConfigInitializer.java
@@ -15,6 +15,7 @@ import ch.qos.logback.classic.joran.JoranConfigurator;
 import ch.qos.logback.classic.util.ContextInitializer;
 import ch.qos.logback.core.joran.spi.JoranException;
 import ch.qos.logback.core.util.StatusPrinter;
+import lombok.experimental.UtilityClass;
 
 /**
  * Initializes logging config using an environment variable. Needed for native
@@ -23,8 +24,10 @@ import ch.qos.logback.core.util.StatusPrinter;
  * @author devaprasadh
  *
  */
+@UtilityClass
 public class LogbackConfigInitializer {
     private static final String ENV_LOGBACK_CONFIG = "LOGBACK_CONFIGURATION_FILE";
+    private static final String PROP_LOGGING_CONFIG = "logger.config";
 
     /**
      * Initializes logging config using an environment variable
@@ -55,7 +58,10 @@ public class LogbackConfigInitializer {
                 context.reset();
                 configurator.doConfigure(config);
             } catch (JoranException exception) {
-                // StatusPrinter will handle this
+                // StatusPrinter will handle logging the exception
+                if (configFile.equals(System.getProperty(PROP_LOGGING_CONFIG))) {
+                    System.clearProperty(PROP_LOGGING_CONFIG);
+                }
             } finally {
                 config.close();
             }
@@ -71,17 +77,24 @@ public class LogbackConfigInitializer {
      */
     private static InputStream readConfig(String configFile) {
         InputStream stream = Application.class.getClassLoader().getResourceAsStream(configFile); // From classpath
+        boolean isNotUrl = true;
 
         if (null == stream) {
             try {
                 stream = new FileInputStream(configFile); // From local file
             } catch (FileNotFoundException exception) {
+                isNotUrl = false;
+
                 try {
                     stream = new URL(configFile).openStream(); // From remote URL
                 } catch (IOException ioException) {
                     // no op
                 }
             }
+        }
+
+        if (isNotUrl) {
+            System.setProperty(PROP_LOGGING_CONFIG, configFile);
         }
 
         return stream;