+++ src/main/java/io/github/devatherock/artifactory/controllers/VersionController.java
@@ -24,7 +24,7 @@ import lombok.extern.slf4j.Slf4j;
  *
  */
 @Slf4j
-@ExecuteOn(TaskExecutors.IO)
+@ExecuteOn(TaskExecutors.BLOCKING)
 @Controller
 @RequiredArgsConstructor
 public class VersionController {