+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - fix(deps): update dependency org.objenesis:objenesis to v3.4
+- fix(deps): update dependency org.wiremock:wiremock to v3.6.0
 
 ## [2.1.0] - 2024-05-07
 ### Added