+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - fix(deps): update dependency org.objenesis:objenesis to v3.4
 - fix(deps): update dependency org.wiremock:wiremock to v3.6.0
+- fix(deps): update dependency org.wiremock:wiremock to v3.7.0
 
 ## [2.1.0] - 2024-05-07
 ### Added