+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fix(deps): update dependency org.objenesis:objenesis to v3.4
 - fix(deps): update dependency org.wiremock:wiremock to v3.6.0
 - fix(deps): update dependency org.wiremock:wiremock to v3.7.0
+- fix(deps): update dependency org.projectlombok:lombok to v1.18.34
 
 ## [2.1.0] - 2024-05-07
 ### Added