+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix(deps): update dependency org.wiremock:wiremock to v3.8.0
 - chore(deps): update plugin org.sonarqube to v5.1.0.4882
 - fix(deps): update dependency org.jsoup:jsoup to v1.18.1
+- fix(deps): update dependency org.wiremock:wiremock to v3.9.1
 
 ## [2.1.0] - 2024-05-07
 ### Added