+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - chore(deps): update plugin org.sonarqube to v5.1.0.4882
 - fix(deps): update dependency org.jsoup:jsoup to v1.18.1
 - fix(deps): update dependency org.wiremock:wiremock to v3.9.1
+- fix(deps): update dependency ch.qos.logback:logback-classic to v1.5.8
 
 ## [2.1.0] - 2024-05-07
 ### Added