+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - fix(deps): update dependency org.jsoup:jsoup to v1.18.1
 - fix(deps): update dependency org.wiremock:wiremock to v3.9.1
 - fix(deps): update dependency ch.qos.logback:logback-classic to v1.5.8
+- fix(deps): update dependency net.logstash.logback:logstash-logback-encoder to v8
 
 ## [2.1.0] - 2024-05-07
 ### Added