+++ CHANGELOG.md
@@ -3,15 +3,13 @@
 ## [Unreleased]
 ### Changed
 - fix(deps): update dependency org.objenesis:objenesis to v3.4
-- fix(deps): update dependency org.wiremock:wiremock to v3.6.0
-- fix(deps): update dependency org.wiremock:wiremock to v3.7.0
 - fix(deps): update dependency org.projectlombok:lombok to v1.18.34
-- fix(deps): update dependency org.wiremock:wiremock to v3.8.0
 - chore(deps): update plugin org.sonarqube to v5.1.0.4882
 - fix(deps): update dependency org.jsoup:jsoup to v1.18.1
 - fix(deps): update dependency org.wiremock:wiremock to v3.9.1
 - fix(deps): update dependency ch.qos.logback:logback-classic to v1.5.8
 - fix(deps): update dependency net.logstash.logback:logstash-logback-encoder to v8
+- Upgraded gradle to `8.10.2`
 
 ## [2.1.0] - 2024-05-07
 ### Added