+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Health check to `docker compose`, to make integration tests wait for app startup
+
 ### Changed
 - fix(deps): update dependency org.objenesis:objenesis to v3.4
 - fix(deps): update dependency org.projectlombok:lombok to v1.18.34