+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Health check to `docker compose`, to make integration tests wait for app startup
+- Tests to verify log format
 
 ### Changed
 - fix(deps): update dependency org.objenesis:objenesis to v3.4