+++ CHANGELOG.md
@@ -15,8 +15,8 @@
 - fix(deps): update dependency net.logstash.logback:logstash-logback-encoder to v8
 - Upgraded gradle to `8.10.2`
 - Switched to `native-image-community` graalvm image
-- fix(deps): update dependency ch.qos.logback:logback-classic to v1.5.9
 - fix(deps): update dependency ch.qos.logback:logback-classic to v1.5.10
+- Upgraded micronaut to `4.6.2`
 
 ## [2.1.0] - 2024-05-07
 ### Added