+++ CHANGELOG.md
@@ -7,12 +7,12 @@
 - chore(deps): update plugin org.sonarqube to v5.1.0.4882
 - fix(deps): update dependency org.jsoup:jsoup to v1.18.1
 - fix(deps): update dependency org.wiremock:wiremock to v3.9.1
-- fix(deps): update dependency ch.qos.logback:logback-classic to v1.5.8
 - fix(deps): update dependency net.logstash.logback:logstash-logback-encoder to v8
 - Upgraded gradle to `8.10.2`
 - Switched to `native-image-community` graalvm image
 - fix(deps): update dependency ch.qos.logback:logback-classic to v1.5.10
 - Upgraded micronaut to `4.6.2`
+- Upgraded jdk to 21
 
 ## [2.1.0] - 2024-05-07
 ### Added